test: make materializer CI coverage hermetic
